Refactored UsbApi
